Moved all files that are on the intern-page to the intern-folder
